modules/image/shape_recognition: fix function casts
The return value of a function should only be casted to a compatible
type; lots of "(int) floor(x)" can be simply changed to the implicit
casting rules.

+++ modules/src/image/shape_recognition/sr_normalize.c
@@ -169,7 +169,7 @@ static int Normalize_affine_step1(Fcurves cs, float *determinant)
     float a,b,c,surface;
     float A,B,C;
     
-    surface = (float)size_fcvs(cs);
+    surface = size_fcvs(cs);
     
     a = moment_pq(cs,2,0);
     b = moment_pq(cs,0,2);
